feat(auto choose class): choose classes for types automatically
